fixed linting issues, reused common type yesNoDontknow, swapped out dob data with generateDOB helper function
