+++ e2e/fixtures/manageCases/caseProgression/amendDetails/amendChildDetails/amendChildDetailsRevised1Content.ts
@@ -24,9 +24,6 @@ export enum AmendChildDetailsRevised1Content {
   formLabelChildGenderOptional = "*Child's gender (Optional)",
   childFirstName = "AmendedFirstName",
   childLastName = "AmendedChildLastName",
-  childDayOfBirth = "1",
-  childMonthOfBirth = "1",
-  childYearOfBirth = "2010",
   optionalChildGender = "Lorem ipsum gender",
   parentalResponsibility = "Child's father and was married to the mother when child was born",
   parentalResponsibility2 = "To the third person",
