--- e2e/journeys/manageCases/caseProgression/amendDetails/amendChildDetails/amendChildDetails.ts
@@ -1,54 +1,0 @@
-import { Browser, Page } from "@playwright/test";
-import { Helpers } from "../../../../../common/helpers.ts";
-import { AmendChildDetailsRevised1Page } from "../../../../../pages/manageCases/caseProgression/amendDetails/amendChildDetails/amendChildDetailsRevised1Page.ts";
-import { AmendChildDetailsRevised2Page } from "../../../../../pages/manageCases/caseProgression/amendDetails/amendChildDetails/amendChildDetailsRevised2Page.ts";
-import { AmendChildDetailsSubmitPage } from "../../../../../pages/manageCases/caseProgression/amendDetails/amendChildDetails/amendChildDetailsSubmitPage.ts";
-import { C100ChildGender } from "../../../../../pages/manageCases/createCase/C100/childDetails/childDetailsRevised1Page.ts";
-import { yesNoDontKnowC100ChildDetailsRevisedAdditionalQuestions } from "../../../../../pages/manageCases/createCase/C100/childDetails/childDetailsRevised2Page.ts";
-import config from "../../../../../config.ts";
-import { DummyC100 } from "../../../createCase/dummyCase/dummyC100.ts";
-
-interface AmendChildDetailsParams {
-  page: Page;
-  accessibilityTest: boolean;
-  c100ChildGender: C100ChildGender;
-  yesNoDontKnowC100ChildDetailsRevisedAdditionalQuestions: yesNoDontKnowC100ChildDetailsRevisedAdditionalQuestions;
-  browser: Browser;
-}
-
-export class AmendChildDetails {
-  public static async amendChildDetails({
-    page,
-    accessibilityTest,
-    c100ChildGender,
-    yesNoDontKnowC100ChildDetailsRevisedAdditionalQuestions,
-    browser,
-}: AmendChildDetailsParams): Promise<void> {
-    const caseRef = await DummyC100.dummyC100({
-      page: page,
-      applicantLivesInRefuge: true,
-      otherPersonLivesInRefuge: false,
-    });
-    // open new browser and sign in as court admin user
-    page = await Helpers.openNewBrowserWindow(browser, "courtAdminStoke");
-    await Helpers.goToCase(page, config.manageCasesBaseURL, caseRef, "tasks");
-
-    await Helpers.chooseEventFromDropdown(page, "Amend Child details");
-    await AmendChildDetailsRevised1Page.amendChildDetailsRevised1Page({
-      page,
-      accessibilityTest,
-      c100ChildGender,
-    });
-    await AmendChildDetailsRevised2Page.amendChildDetailsRevised2Page({
-      page,
-      accessibilityTest,
-      yesNoDontKnowC100ChildDetailsRevisedAdditionalQuestions,
-    });
-    await AmendChildDetailsSubmitPage.amendChildDetailsSubmitPage({
-      page,
-      accessibilityTest,
-      c100ChildGender,
-      yesNoDontKnowC100ChildDetailsRevisedAdditionalQuestions,
-    });
-  }
-}
